validate json config whenever it is opened in browser
