+++ lib/web.js
@@ -6,14 +6,17 @@
 const utils = require('@iobroker/adapter-core'); // Get common adapter utils
 const IoBWebServer = require('@iobroker/webserver');
 const express = require('express');
-const fs = require('fs');
+const fs = require('node:fs');
 const util = require('util');
-const path = require('path');
+const path = require('node:path');
 const stream = require('stream');
 const compression = require('compression');
 const mime = require('mime');
 const zlib = require('zlib');
 const archiver = require('archiver');
+const { default: axios } = require('axios');
+const Ajv = require('ajv');
+const JSON5 = require('json5');
 
 let session;
 let bodyParser;
@@ -143,6 +146,9 @@ class Web {
 
     // todo delete after react will be main
     LOGIN_PAGE = '/index.html?login';
+
+    /** URL to the JSON config schema */
+    JSON_CONFIG_SCHEMA_URL = 'https://raw.githubusercontent.com/ioBroker/adapter-react-v5/main/schemas/jsonConfig.json';
 
     bruteForce = {};
     store = null;
@@ -345,6 +351,51 @@ class Web {
             }
         } else {
             return `${this.LOGIN_PAGE}?error`;
+        }
+    }
+
+    /**
+     * Validate, al JSON configs from alla adapters against the current schema
+     *
+     * @param {string} adapterName name of the adapter
+     */
+    async validateJsonConfig(adapterName) {
+        const schemaRes = await axios.get(this.JSON_CONFIG_SCHEMA_URL);
+        const schema = schemaRes.data;
+
+        const res = await this.adapter.getForeignObjectAsync(`system.adapter.${adapterName}`);
+
+        if (res?.common.adminUI?.config === 'json') {
+            try {
+                const ajv = new Ajv({ allErrors: false });
+
+                const adapterPath = path.dirname(require.resolve(`iobroker.${adapterName}/package.json`));
+
+                const jsonConfPath = path.join(adapterPath, 'admin', 'jsonConfig.json');
+                const json5ConfPath = path.join(adapterPath, 'admin', 'jsonConfig.json5');
+                let jsonConf;
+
+                if (fs.existsSync(jsonConfPath)) {
+                    jsonConf = fs.readFileSync(jsonConfPath, {
+                        encoding: 'utf-8',
+                    });
+                } else {
+                    jsonConf = fs.readFileSync(json5ConfPath, {
+                        encoding: 'utf-8',
+                    });
+                }
+
+                const validate = ajv.compile(schema);
+                const valid = validate(JSON5.parse(jsonConf));
+
+                if (!valid) {
+                    this.adapter.log.warn(
+                        `${adapterName} has an invalid jsonConfig: ${JSON.stringify(validate.errors)}`
+                    );
+                }
+            } catch (e) {
+                this.adapter.log.debug(`Error validating schema of ${adapterName}: ${e.message}`);
+            }
         }
     }
 
@@ -670,6 +721,14 @@ class Web {
             }
 
             this.server.app.get('/iobroker_check.html', (req, res) => res.send('ioBroker'));
+
+            this.server.app.get('/validate_config/*', async (req, res) => {
+                const adapterName = req.url.split('/').pop();
+
+                await this.validateJsonConfig(adapterName.toLowerCase());
+
+                res.status(200).send('validated');
+            });
 
             this.server.app.get('/zip/*', (req, res) => {
                 const parts = req.url.split('/');
@@ -1225,7 +1284,11 @@ class Web {
                                         this.checkTimeout = setTimeout(async () => {
                                             this.checkTimeout = null;
                                             try {
-                                                await IoBWebServer.checkPublicIP(this.settings.port, 'ioBroker', '/iobroker_check.html');
+                                                await IoBWebServer.checkPublicIP(
+                                                    this.settings.port,
+                                                    'ioBroker',
+                                                    '/iobroker_check.html'
+                                                );
                                             } catch (e) {
                                                 // this supported first from js-controller 5.0.
                                                 this.adapter.sendToHost(
